[KERNEL/USER] Read/write request offset member is now an uint64_t

+++ include/kernel/vfs/message.h
@@ -91,7 +91,7 @@ struct req_rdwr {
 
     size_t size;
 
-    size_t off;
+    uint64_t off;
 
     void *data;
 };
